requirements.txt
django==1.10
dj-database-url==0.4
djangorestframework==3.4
requests==2.10
psycopg2==2.6
PyJWT==1.4.1